requirements.txt
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
pymongo>=4.6.3
redis>=5.0.1
qdrant-client==1.11.0
neo4j>=5.17.0
pika>=1.3.2
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
